修改联系人表、负责人表的存储结构，重构相关代码；实现第三方的相关功能

+++ src/main/java/com/glimound/rmsbackend/pojo/ResearchAchievement.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ResearchAchievement {
-    private Long achievementId;
+    private Integer achievementId;
     private String achievementName;
     private LocalDate acquisitionDate;
     private Integer achievementType;
